requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohappyeyeballs==2.5.0
aiohttp==3.10.11
aiosignal==1.3.2
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arviz==0.21.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
babel==2.17.0
backtesting==0.6.2
beautifulsoup4==4.13.3
bleach==6.2.0
bokeh==3.6.3
build==0.10.0
CacheControl==0.12.14
cachetools==5.5.0
ccxt==4.4.65
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.0
cloudpickle==3.1.1
comm==0.2.2
conda-pack==0.8.1
cons==0.4.6
contourpy==1.3.1
crashtest==0.4.1
cryptography==44.0.2
cycler==0.12.1
dateparser==1.2.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.9
dulwich==0.21.7
etuples==0.3.9
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
gast==0.6.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5netcdf==1.6.1
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
installer==0.6.0
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.5
isoduration==20.11.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.9.0
keyring==23.13.1
kiwisolver==1.4.8
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.39.1
logical-unification==0.4.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
miniKanren==1.0.3
mistune==3.1.2
ml-dtypes==0.4.1
multidict==6.1.0
multipledispatch==1.0.0
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.0.2
nvidia-nccl-cu12==2.25.1
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.14.1
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandas_ta==0.3.14b0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
pkginfo==1.11.2
platformdirs==2.6.2
plotly==5.24.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycares==4.5.0
pycparser==2.22
Pygments==2.19.1
pymc==5.21.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
pytensor==2.28.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
ta==0.11.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.0.2
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2024.10.21.16
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.3.0
virtualenv==20.21.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xarray==2025.1.2
xarray-einstats==0.8.0
xgboost==2.1.4
xyzservices==2025.1.0
yarl==1.18.3
yfinance==0.2.54
zipp==3.20.2
